improve: compare strings to strings instead of len
It is a lot more obvious to a reader when comparing strings to "" then
a check of the length of the string.

Also rename a few things in an attempt to make them easier to understand.

Rename redirects to renames, since these migrations don't really work
for the general case of redirects. They only work when changing the path
of a route.